implemented logic so user can only make operations in his own notes

+++ backend/src/controllers/notes.ts
@@ -5,10 +5,14 @@ import UpdateNoteBody from '../interfaces/UpdateNoteBody';
 import UpdateNoteParams from '../interfaces/UpdateNoteParams';
 import createHttpError from 'http-errors';
 import mongoose from 'mongoose';
+import { assertIsDefined } from '../utils/assertIsDefined';
 
 const getNotes: RequestHandler = async (req, res, next) => {
+	const authenticatedUserId = req.session.userId;
+
 	try {
-		const notes = await NoteModel.find().exec();
+		assertIsDefined(authenticatedUserId);
+		const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
 		res.status(200).json(notes);
 	} catch (error){
 		next(error);
@@ -17,7 +21,11 @@ const getNotes: RequestHandler = async (req, res, next) => {
 
 const getNote: RequestHandler = async (req, res, next) => {
 	const { id } = req.params;
+	const authenticatedUserId = req.session.userId;
+
 	try {
+		assertIsDefined(authenticatedUserId);
+
 		if (!mongoose.isValidObjectId(id)) {
 			throw createHttpError(400, 'Invalid note ID format');
 		}
@@ -28,6 +36,10 @@ const getNote: RequestHandler = async (req, res, next) => {
 			throw createHttpError(404, 'Note not found');
 		}
 
+		if(!note.userId.equals(authenticatedUserId)) {
+			throw createHttpError(401, 'You cannot access this note');
+		}
+
 		res.status(200).json(note);
 	} catch (error) {
 		next(error);
@@ -36,13 +48,17 @@ const getNote: RequestHandler = async (req, res, next) => {
 
 const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
 	const { title, text } = req.body;
+	const authenticatedUserId = req.session.userId;
 
 	try {
+		assertIsDefined(authenticatedUserId);
+
 		if (!title) {
 			throw createHttpError(400, 'Note must have a title');
 		}
 
 		const newNote = await NoteModel.create({
+			userId: authenticatedUserId,
 			title,
 			text,
 		});
@@ -55,8 +71,12 @@ const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = as
 const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBody, unknown> = async (req, res, next) => {
 	const { id } = req.params;
 	const { title, text } = req.body;
+	const authenticatedUserId = req.session.userId;
+
 
 	try {
+		assertIsDefined(authenticatedUserId);
+
 		if (!mongoose.isValidObjectId(id)) {
 			throw createHttpError(400, 'Invalid note ID format');
 		}
@@ -69,6 +89,10 @@ const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBody, unkn
 
 		if (!note) {
 			throw createHttpError(404, 'Note not found');
+		}
+
+		if(!note.userId.equals(authenticatedUserId)) {
+			throw createHttpError(401, 'You cannot access this note');
 		}
 
 		note.title = title;
@@ -84,8 +108,11 @@ const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBody, unkn
 
 const deleteNote: RequestHandler = async (req, res, next) => {
 	const { id } = req.params;
+	const authenticatedUserId = req.session.userId;
 
 	try {
+		assertIsDefined(authenticatedUserId);
+
 		if (!mongoose.isValidObjectId(id)) {
 			throw createHttpError(400, 'Invalid note ID format');
 		}
@@ -94,6 +121,11 @@ const deleteNote: RequestHandler = async (req, res, next) => {
 
 		if (!note) {
 			throw createHttpError(404, 'Note not found');
+		}
+
+
+		if(!note.userId.equals(authenticatedUserId)) {
+			throw createHttpError(401, 'You cannot access this note');
 		}
 
 		await note.deleteOne();
